Excel version of dict.textile
This is an Excel version of the `dict.textile` file that is queried from the git source using an ODBC query inside Excel. Each letter gets its own sheet. This file is dynamically populated by Excel and its contents should not be edited directly. It is primarily for printing purposes.